Removed async task to retrieve zoo schedule for the day. Added the logic to HomePageViewModel.
