+++ app/src/main/java/ca/judacribz/week2weekend/homepage/model/Schedule.kt
@@ -1,0 +1,3 @@
+package ca.judacribz.week2weekend.homepage.model
+
+data class Schedule(val admissionTime: String, val lastAdmission: String)
